Thanks to Geoff Norton for tracking the last major Swipe problem down.
It turns out that adding our "menuView" to the cell.ContentView caused
the menuView.BackgroundColor to be changed by the UITableView code based
on the selection status (they insert a blue view in the background and
probably change all the subviews' BackgroundColor of the view to be
transparent).

This rendered the menu white sometimes with no pattern.

In addition now the texture color is now cached.

The menu is cancelled when new data is loaded to ensure that the menu
is not up when we insert cells.