Introducting TrueFeature(tm)
TrueFeature(tm) is a feature that was requested by a RealUser(tm)
on Twitter(tm).    This feature was requested by @jacksonh and is
pure brilliance.

Whenever you retweet a tweet, it becomes automatically favorited.
For the lagards among you, I have added a setting so you can turn
this off.